+++ NEWS.md
@@ -1,3 +1,23 @@
+# bakR 0.4.0
+* Added QC_check(), a function to perform quality control analysis with bakRFit objects. Looks for any problems in your data that will impair bakR's performance, generates a number of diagnostic visualizations, and makes suggestions about what to do next.
+* Fixed plot coloring bug in plotMA() and plotVolcano().
+* Fixed bug that led to problems when the number of -s4U replicates > +s4U replicates in one or more Exp_IDs
+* Implemented improved U-content adjustment for MCMC implementation. Also impacts accuracy of StanRateEst = TRUE mutation rate estimation strategy.
+* Increased default number of features to use for StanRateEst mutation rate estimation strategy.
+* Improved scaling of NSSHeat() output matrix columns.
+* Created a new function NSSHeat2() that implements a different mechanism scoring function than NSSHeat().
+
+# bakR 0.3.0
+
+* Optimized data preprocessing with data.table
+* Increased simulation flexibility
+* Added edge case error catching
+
+# bakR 0.2.5
+
+* Added checks to bakRData validator
+* Fixed bug in NSSHeatmap that prevented adjusting padj cutoff
+
 # bakR 0.2.4
 
 * Addressed NOTEs to prepare for CRAN submission